Add line break and spacing for daily forecast.
This commit aims to make the cli more friendly for smaller
monitors and split screens.

The PrintDaily() function's output produces over 80 characters in a
single line. To fix this, the Printf() function in PrintDaily()
is split into two to allow for a line break.

All Printf()'s in printCommon() are moved over 2 spaces to make
some visual breaks in the output. It's necessary because PrintDaily()
now adds an extra line to the forecast output and doesn't look very
pretty with forecasts that are over 2 days long.