Remove "Day Forecast" title and fix line spacing
The philosophy around this commit is that displaying the amount
of forecasted days a user has defined with the -l flag isn't that
meaningful in the context of use from a terminal and it breaks the
visual flow of the output.

When I removed the title in the println(), I ran into some spacing
issues and found that the line break in printDaily() located right before
printing the date works much nicer if moved to printCommon() after
the printing of the air pressure. An added bonus to moving the line
break is that there is now spacing between the last line of output and
the terminal prompt.

I think that the visual flow is much nicer now when using the -l flag.