Improvements in formatting for multi-day forecasts (#33)
* Add line break and spacing for daily forecast.

This commit aims to make the cli more friendly for smaller
monitors and split screens.

The PrintDaily() function's output produces over 80 characters in a
single line. To fix this, the Printf() function in PrintDaily()
is split into two to allow for a line break.

All Printf()'s in printCommon() are moved over 2 spaces to make
some visual breaks in the output. It's necessary because PrintDaily()
now adds an extra line to the forecast output and doesn't look very
pretty with forecasts that are over 2 days long.

* Remove "Day Forecast" title and fix line spacing

The philosophy around this commit is that displaying the amount
of forecasted days a user has defined with the -l flag isn't that
meaningful in the context of use from a terminal and it breaks the
visual flow of the output.

When I removed the title in the println(), I ran into some spacing
issues and found that the line break in printDaily() located right before
printing the date works much nicer if moved to printCommon() after
the printing of the air pressure. An added bonus to moving the line
break is that there is now spacing between the last line of output and
the terminal prompt.

I think that the visual flow is much nicer now when using the -l flag.